requirements.txt
# Python version 3.11.1
aiolimiter==1.1.0
anyio==3.7.1
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gspread==5.10.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
oauth2client==4.1.3
oauthlib==3.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyparsing==3.1.0
python-telegram-bot==13.13
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
tornado==6.3.2
tzlocal==5.0.1
urllib3==1.26.16
